Rename resource card for convention
